Moves git related files into tracker
Preparing to make my git settings portable to others, so all
settings that are local to me (such as name, email) are now stored in
gitconfig.local. A separate git directory is also available which will
potentially hold all file settings that were previously in gitconfig
sections. This would make it easier to contextualize for anyone wanting
to know what were my gitconfig aliases for example.