Adds local to .gitignore
Any directory named local will automatically be ignored by git. This
line of thinking separates out the stuff that is local to the
machine/operating system/environment/user/company from what is
publishable. Sensitive content can go in this directory keeping in mind
that it wont be backed up or tracked.