Fixes bug: git status does not show one change
Bugfix for parse-git-dirty in situaions where the change was just a
single file. Previously, git status --porcelain returned incorrect
dirty status because /usr/bin/time returned incorrect results on
mounted systems (using git from within a VM and accessing a repo on
a Windows share).

More testing might be needed but hopefully, that situation is mitigated
now with this bugfix.